tests(web): fix test mocks for web3 module split
Update test files to mock web3ReadOnly module instead of web3 module
after the module was split. The source files now import hooks from
web3ReadOnly.ts, so tests need to mock that module for spies to work.
