+++ apps/web/src/components/new-safe/create/__tests__/useEstimateSafeCreationGas.test.ts
@@ -3,7 +3,7 @@ import { useEstimateSafeCreationGas } from '@/components/new-safe/create/useEsti
 import * as chainIdModule from '@/hooks/useChainId'
 import { type ConnectedWallet } from '@/hooks/wallets/useOnboard'
 import * as wallet from '@/hooks/wallets/useWallet'
-import * as web3 from '@/hooks/wallets/web3'
+import * as web3ReadOnly from '@/hooks/wallets/web3ReadOnly'
 import * as safeContracts from '@/services/contracts/safeContracts'
 import * as store from '@/store'
 import { renderHook } from '@/tests/test-utils'
@@ -62,7 +62,7 @@ describe('useEstimateSafeCreationGas', () => {
 
   it('should estimate gas', async () => {
     const mockProvider = new JsonRpcProvider()
-    jest.spyOn(web3, 'useWeb3ReadOnly').mockReturnValue(mockProvider)
+    jest.spyOn(web3ReadOnly, 'useWeb3ReadOnly').mockReturnValue(mockProvider)
     jest.spyOn(sender, 'estimateSafeCreationGas').mockReturnValue(Promise.resolve(BigInt('123')))
     jest.spyOn(wallet, 'default').mockReturnValue({
       label: 'MetaMask',
@@ -90,7 +90,7 @@ describe('useEstimateSafeCreationGas', () => {
   })
 
   it('should not estimate gas if there is no provider', async () => {
-    jest.spyOn(web3, 'useWeb3ReadOnly').mockReturnValue(undefined)
+    jest.spyOn(web3ReadOnly, 'useWeb3ReadOnly').mockReturnValue(undefined)
     const { result } = renderHook(() => useEstimateSafeCreationGas(mockProps))
 
     await waitFor(() => {
