+++ apps/web/src/components/tx/security/tenderly/__tests__/utils.test.ts
@@ -8,7 +8,7 @@ import EthSafeTransaction from '@safe-global/protocol-kit/dist/src/utils/transac
 import { ZERO_ADDRESS } from '@safe-global/protocol-kit/dist/src/utils/constants'
 import { generatePreValidatedSignature } from '@safe-global/protocol-kit/dist/src/utils/signatures'
 import { toBeHex } from 'ethers'
-import * as Web3 from '@/hooks/wallets/web3'
+import * as Web3 from '@/hooks/wallets/web3ReadOnly'
 import {
   NONCE_STORAGE_POSITION,
   THRESHOLD_STORAGE_POSITION,
